+++ Sources/BatteryPanicApp/Shared/AppConstants.swift
@@ -4,5 +4,6 @@ enum AppConstants {
     static let appName = "Battery Panic"
     static let defaultThreshold = 10
     static let pollInterval: TimeInterval = 5
-    static let defaultPreviewDuration: TimeInterval = 8
+    static let previewAlarmDuration: TimeInterval = 4
+    static let warningSoundRepeatInterval: TimeInterval = 10
 }